Com_05
1. Create HealthComponent: Added to owner HP, event Take Damdge, event Die, event AddHealth.
2. Modyfire Enemy, Add HealthComponent.
3. Modyfire VRPavn. Add HealthComponent.
4. Modyfire KillVoluem. Added connection for work HealthComponent.

+++ Source/VRGamePreview/Item/MainTypes.h
@@ -53,9 +53,9 @@ struct FProjectileInfo
 	UNiagaraSystem* TrailFXNi = nullptr;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ProjectileSetting")
-	float ProjectileMinDamage = 20.0f;
+	int ProjectileMinDamage = 20;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ProjectileSetting")
-	float ProjectileMaxDamage = 20.0f;
+	int ProjectileMaxDamage = 20;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ProjectileSetting")
 	float ProjectileLifeTime = 20.0f;
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "ProjectileSetting")
